jobs/llm_finetuning.py
"""LLM Fine-Tuning Job Execution"""

import torch
from transformers import AutoModelForCausalLM, AutoTokenizer, Trainer, TrainingArguments
from peft import LoraConfig, get_peft_model
from typing import Dict, Any
import structlog

from .base import BaseJob

log = structlog.get_logger()


class LLMFineTuningJob(BaseJob):
    """Execute LLM fine-tuning jobs with LoRA"""

    def __init__(self, job_data: Dict[str, Any]):
        super().__init__(job_data)
        self.model_name = self.input_data.get('model', 'Qwen/Qwen-7B')
        self.dataset_url = self.input_data.get('dataset_url', '')
        self.epochs = self.config.get('epochs', 3)
        self.learning_rate = self.config.get('learning_rate', 2e-4)
        self.lora_r = self.config.get('lora_r', 8)
        self.lora_alpha = self.config.get('lora_alpha', 16)

    def validate(self) -> bool:
        if not super().validate():
            return False

        if not self.dataset_url:
            log.error("Dataset URL is required for fine-tuning")
            return False

        return True

    def execute(self) -> Dict[str, Any]:
        """Execute fine-tuning with LoRA"""
        try:
            if not self.validate():
                return {'success': False, 'error': 'Validation failed'}

            log.info("Starting fine-tuning", model=self.model_name, epochs=self.epochs)

            # Load model
            model = AutoModelForCausalLM.from_pretrained(
                self.model_name,
                torch_dtype=torch.float16,
                device_map="auto",
                trust_remote_code=True
            )

            # Apply LoRA
            lora_config = LoraConfig(
                r=self.lora_r,
                lora_alpha=self.lora_alpha,
                target_modules=["q_proj", "v_proj"],
                lora_dropout=0.05,
                bias="none",
                task_type="CAUSAL_LM"
            )

            model = get_peft_model(model, lora_config)
            model.print_trainable_parameters()

            # TODO: Load and preprocess dataset from dataset_url
            # TODO: Setup Trainer with training_args
            # TODO: Execute training
            # TODO: Save adapter weights
            # TODO: Upload to IPFS/S3

            log.info("Fine-tuning completed")

            return {
                'success': True,
                'message': 'Fine-tuning completed',
                'model': self.model_name,
                'checkpoint_url': 'ipfs://...'
            }

        except Exception as e:
            log.error("Fine-tuning failed", error=str(e))
            return {
                'success': False,
                'error': str(e)
            }

    def cleanup(self):
        """Clean up"""
        if torch.cuda.is_available():
            torch.cuda.empty_cache()
        log.info("Cleanup completed")
