requirements.txt
# Core dependencies
requests==2.31.0
websocket-client==1.7.0
pydantic==2.5.3
python-dotenv==1.0.0

# GPU detection
py3nvml==0.2.7
pynvml==11.5.0
GPUtil==1.4.0

# Job execution
torch==2.1.2
transformers==4.36.2
accelerate==0.25.0
bitsandbytes==0.41.3

# LLM support
ollama==0.1.6
openai==1.6.1

# Vector databases (for RAG)
faiss-cpu==1.7.4
chromadb==0.4.22
sentence-transformers==2.2.2

# Utils
psutil==5.9.7
structlog==24.1.0
aiohttp==3.9.1
tenacity==8.2.3
